# examples/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "ListUsers",
        "description": "ListUsers handles listing all users.",
        "tags": [
          "users"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "operationId": "CreateUser",
        "description": "CreateUser handles user creation requests.",
        "tags": [
          "users"
        ],
        "security": [
          {
            "BasicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "return-to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminUserResponse": {
        "type": "object",
        "properties": {
          "UserResponse": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "description": "AdminUserResponse represents an admin user with additional fields."
      },
...
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username"
        ],
        "description": "CreateUserRequest represents the request body for creating a user."
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "details": {
            "type": "object"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "ListUsersResponse": {
        "oneOf": [
          {
            "type": "array",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/AdminUserResponse"
              }
            }
          },
          {
            "type": "array",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/UserResponse"
              }
            }
          }
        ],
        "description": "ListUsersResponse represents the response for listing users."
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "userID": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "description": "UserResponse represents the response for user operations."
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "details": {
            "type": "object"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Unprocessable Entity - Request body could not be parsed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}